Added support for ImageSource based MaterialUnderlinedTabItem
