--- Tabs/Tabs/ImageIconUnderlinedTabItem.xaml
@@ -1,74 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8" ?>
-
-<tabs:UnderlinedTabItemBase 
-    x:Class="Sharpnado.Tabs.ImageIconUnderlinedTabItem"
-    xmlns="http://xamarin.com/schemas/2014/forms"
-    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-    xmlns:tabs="clr-namespace:Sharpnado.Tabs"
-    x:Name="RootLayout">
-
-    <ContentView.Resources>
-        <ResourceDictionary>
-            <Color x:Key="HeaderTextColor">#acacac</Color>
-
-            <Style x:Key="TabIcon" TargetType="Image">
-                <Setter Property="VerticalOptions" Value="End" />
-                <Setter Property="HorizontalOptions" Value="Center" />
-            </Style>
-
-            <Style x:Key="TabTextHeader" TargetType="Label">
-                <Setter Property="TextColor" Value="{StaticResource HeaderTextColor}" />
-                <Setter Property="LineBreakMode" Value="WordWrap" />
-                <Setter Property="HorizontalTextAlignment" Value="Center" />
-                <Setter Property="VerticalTextAlignment" Value="Center" />
-            </Style>
-        </ResourceDictionary>
-    </ContentView.Resources>
-
-    <ContentView.Content>
-        <Grid
-            x:Name="Grid"
-            BackgroundColor="Transparent"
-            ColumnSpacing="0"
-            RowSpacing="0">
-
-            <Grid.RowDefinitions>
-                <RowDefinition Height="8*" />
-                <RowDefinition x:Name="TextRowDefinition" Height="5*" />
-            </Grid.RowDefinitions>
-
-            <Image
-                x:Name="Icon"
-                HeightRequest="{Binding Source={x:Reference RootLayout}, Path=IconSize}"
-                Source="{Binding Source={x:Reference RootLayout}, Path=IconImageSource}"
-                Style="{StaticResource TabIcon}"
-                WidthRequest="{Binding Source={x:Reference RootLayout}, Path=IconSize}" />
-
-            <BoxView
-                x:Name="Underline"
-                Grid.Row="1"
-                Margin="0,0"
-                BackgroundColor="Transparent"
-                HeightRequest="0"
-                HorizontalOptions="Center"
-                VerticalOptions="End">
-                <BoxView.Triggers>
-                    <DataTrigger
-                        Binding="{Binding Source={x:Reference RootLayout}, Path=IsSelected}"
-                        TargetType="BoxView"
-                        Value="true">
-                        <Setter Property="HeightRequest" Value="{Binding Source={x:Reference RootLayout}, Path=UnderlineHeight}" />
-                    </DataTrigger>
-                </BoxView.Triggers>
-            </BoxView>
-            <Label
-                x:Name="InnerLabel"
-                Grid.Row="1"
-                FontFamily="{Binding Source={x:Reference RootLayout}, Path=FontFamily}"
-                FontSize="{Binding Source={x:Reference RootLayout}, Path=LabelSize}"
-                HorizontalOptions="Center"
-                Style="{StaticResource TabTextHeader}"
-                Text="{Binding Source={x:Reference RootLayout}, Path=Label}" />
-        </Grid>
-    </ContentView.Content>
-</tabs:UnderlinedTabItemBase>
